+++ ooniprobe/AppDelegate.mm
@@ -1,7 +1,3 @@
-// Part of MeasurementKit <https://measurement-kit.github.io/>.
-// MeasurementKit is free software. See AUTHORS and LICENSE for more
-// information on the copying conditions.
-
 #import "AppDelegate.h"
 #import "Tests.h"
 #import "NotificationService.h"
@@ -9,8 +5,8 @@
 #import "DictionaryUtility.h"
 #import "RunTestViewController.h"
 
-#define alert_tag_notification 1
-#define alert_tag_open_itunes 2
+#define ALERT_TAG_NOTIFICATION 1
+#define ALERT_TAG_OPEN_ITUNES 2
 
 @interface AppDelegate ()
 
@@ -61,8 +57,11 @@
 - (void)application:(UIApplication *)application didReceiveRemoteNotification:(NSDictionary *)userInfo {
     if (userInfo) {
         if ([userInfo objectForKey:@"aps"]){
-            if([[userInfo objectForKey:@"aps"] objectForKey:@"badge"])
-                [UIApplication sharedApplication].applicationIconBadgeNumber = [[[userInfo objectForKey:@"aps"] objectForKey: @"badge"] intValue];
+            if([[userInfo objectForKey:@"aps"] objectForKey:@"badge"]){
+                dispatch_async(dispatch_get_main_queue(), ^{
+                    [UIApplication sharedApplication].applicationIconBadgeNumber = [[[userInfo objectForKey:@"aps"] objectForKey: @"badge"] intValue];
+                });
+            }
         }
         [self handleNotification:userInfo :application];
     }
@@ -82,7 +81,7 @@
                 NSArray *alt_href = [[userInfo objectForKey:@"payload"] objectForKey:@"alt_hrefs"];
                 [links addObjectsFromArray:alt_href];
             }
-            alertView.tag = alert_tag_notification;
+            alertView.tag = ALERT_TAG_NOTIFICATION;
             [alertView show];
         }
         else {
@@ -117,12 +116,12 @@
 
 - (void)alertView:(UIAlertView *)alertView clickedButtonAtIndex:(NSInteger)buttonIndex
 {
-    if (buttonIndex == 1 && alertView.tag == alert_tag_notification){
+    if (buttonIndex == 1 && alertView.tag == ALERT_TAG_NOTIFICATION){
         [self openBrowser];
     }
-    else if (buttonIndex == 1 && alertView.tag == alert_tag_open_itunes){
+    else if (buttonIndex == 1 && alertView.tag == ALERT_TAG_OPEN_ITUNES){
         //open ooniprobe on iTunes connect
-        NSString *iTunesLink = @"itms://itunes.apple.com/us/app/apple-store/id1245670385?mt=8";
+        NSString *iTunesLink = @"itms://itunes.apple.com/us/app/apple-store/id1199566366?mt=8";
         [[UIApplication sharedApplication] openURL:[NSURL URLWithString:iTunesLink]];
     }
 }
@@ -143,7 +142,9 @@
 
 - (void)applicationDidBecomeActive:(UIApplication *)application {
     // Restart any tasks that were paused (or not yet started) while the application was inactive. If the application was previously in the background, optionally refresh the user interface.
-    [UIApplication sharedApplication].applicationIconBadgeNumber = 0;
+    dispatch_async(dispatch_get_main_queue(), ^{
+        [UIApplication sharedApplication].applicationIconBadgeNumber = 0;
+    });
 }
 
 - (void)applicationWillTerminate:(UIApplication *)application {
@@ -187,7 +188,7 @@
         if ([minimum_version compare:[[[NSBundle mainBundle] infoDictionary] objectForKey:@"CFBundleShortVersionString"] options:NSNumericSearch] == NSOrderedDescending) {
             //actualVersion is lower than the requiredVersion
             UIAlertView *alertView = [[UIAlertView alloc] initWithTitle:NSLocalizedString(@"ooniprobe_outdate", nil) message:NSLocalizedString(@"ooniprobe_outdate_msg", nil) delegate:nil cancelButtonTitle:NSLocalizedString(@"Cancel", nil) otherButtonTitles:NSLocalizedString(@"ok", nil), nil];
-            alertView.tag = alert_tag_open_itunes;
+            alertView.tag = ALERT_TAG_OPEN_ITUNES;
             [alertView show];
         }
         else {