+++ ooniprobe/Utility/VersionUtility.mm
@@ -1,11 +1,3 @@
-//
-//  VersionUtility.m
-//  ooniprobe
-//
-//  Created by Lorenzo Primiterra on 24/09/17.
-//  Copyright © 2017 Simone Basso. All rights reserved.
-//
-
 #import "VersionUtility.h"
 
 @implementation VersionUtility
@@ -13,11 +5,11 @@
 + (NSString*)get_software_version{
     // FIXME: this breaks OONI backend so we don't send the build number
     /*
-    return [NSString stringWithFormat:@"%@%@+%@", [[[NSBundle mainBundle] infoDictionary] objectForKey:@"CFBundleShortVersionString"], release_name, [[[NSBundle mainBundle] infoDictionary] objectForKey:@"CFBundleVersion"]];
+    return [NSString stringWithFormat:@"%@%@+%@", [[[NSBundle mainBundle] infoDictionary] objectForKey:@"CFBundleShortVersionString"], RELEASE_NAME, [[[NSBundle mainBundle] infoDictionary] objectForKey:@"CFBundleVersion"]];
      */
     return [NSString stringWithFormat:@"%@%@",
             [[[NSBundle mainBundle] infoDictionary] objectForKey:@"CFBundleShortVersionString"],
-            release_name];
+            RELEASE_NAME];
 
 }
 @end