+++ ChangeLog.md
@@ -1,4 +1,4 @@
-# ooniprobe-ios 1.2.1 [2017-09-25]
+# ooniprobe-ios 1.2.2 [2017-09-27]
 
 New features:
 
@@ -8,11 +8,25 @@ New features:
 
 * Drop support for iOS < 9
 
-measurement-kit version: 0.7.3
+measurement-kit version: 0.7.4
 
 ooniprobe-wui version: 2.2.4
 
-# ooniprobe-ios 1.2.0 [version yanked]
+# ooniprobe-ios 1.2.2-rc.1
+
+* Fixed max_runtime being used when running from URI scheme
+
+measurement-kit version: 0.7.4
+
+ooniprobe-wui version: 2.2.4
+
+# ooniprobe-ios 1.2.0-rc.2 & 1.2.0-rc.3 & 1.2.1 [version yanked]
+
+measurement-kit version: 0.7.4
+
+ooniprobe-wui version: 2.2.4
+
+# ooniprobe-ios 1.2.0-rc.1 [version yanked]
 
 measurement-kit version: 0.7.2